requirements.txt
django==5.0.6
gunicorn==21.2.0
whitenoise==6.6.0
psycopg[binary]==3.1.18
python-dotenv==1.0.1
markdown2==2.4.13
django-filter==24.2
dj-database-url==2.1.0
pytest==8.2.2
pytest-django==4.9.0
coverage==7.5.4
